@@ -9,11 +9,12 @@ python3 tools/codespell_ignore_lines_from_errors.py /tmp/codespell_errors.txt > 
 git diff to review changes, then commit, push.
 """
 
+from __future__ import annotations
+
 import sys
-from typing import List
 
 
-def run(args: List[str]) -> None:
+def run(args: list[str]) -> None:
     assert len(args) == 1, "codespell_errors.txt"
     cache = {}
     done = set()
