@@ -15,37 +15,26 @@ TEST_SUBMODULE(docstring_options, m) {
         py::options options;
         options.disable_function_signatures();
 
-        m.def(
-            "test_function1", [](int, int) {}, py::arg("a"), py::arg("b"));
-        m.def(
-            "test_function2", [](int, int) {}, py::arg("a"), py::arg("b"), "A custom docstring");
+        m.def("test_function1", [](int, int) {}, py::arg("a"), py::arg("b"));
+        m.def("test_function2", [](int, int) {}, py::arg("a"), py::arg("b"), "A custom docstring");
 
-        m.def(
-            "test_overloaded1", [](int) {}, py::arg("i"), "Overload docstring");
-        m.def(
-            "test_overloaded1", [](double) {}, py::arg("d"));
+        m.def("test_overloaded1", [](int) {}, py::arg("i"), "Overload docstring");
+        m.def("test_overloaded1", [](double) {}, py::arg("d"));
 
-        m.def(
-            "test_overloaded2", [](int) {}, py::arg("i"), "overload docstring 1");
-        m.def(
-            "test_overloaded2", [](double) {}, py::arg("d"), "overload docstring 2");
+        m.def("test_overloaded2", [](int) {}, py::arg("i"), "overload docstring 1");
+        m.def("test_overloaded2", [](double) {}, py::arg("d"), "overload docstring 2");
 
-        m.def(
-            "test_overloaded3", [](int) {}, py::arg("i"));
-        m.def(
-            "test_overloaded3", [](double) {}, py::arg("d"), "Overload docstr");
+        m.def("test_overloaded3", [](int) {}, py::arg("i"));
+        m.def("test_overloaded3", [](double) {}, py::arg("d"), "Overload docstr");
 
         options.enable_function_signatures();
 
-        m.def(
-            "test_function3", [](int, int) {}, py::arg("a"), py::arg("b"));
-        m.def(
-            "test_function4", [](int, int) {}, py::arg("a"), py::arg("b"), "A custom docstring");
+        m.def("test_function3", [](int, int) {}, py::arg("a"), py::arg("b"));
+        m.def("test_function4", [](int, int) {}, py::arg("a"), py::arg("b"), "A custom docstring");
 
         options.disable_function_signatures().disable_user_defined_docstrings();
 
-        m.def(
-            "test_function5", [](int, int) {}, py::arg("a"), py::arg("b"), "A custom docstring");
+        m.def("test_function5", [](int, int) {}, py::arg("a"), py::arg("b"), "A custom docstring");
 
         {
             py::options nested_options;
@@ -59,8 +48,7 @@ TEST_SUBMODULE(docstring_options, m) {
         }
     }
 
-    m.def(
-        "test_function7", [](int, int) {}, py::arg("a"), py::arg("b"), "A custom docstring");
+    m.def("test_function7", [](int, int) {}, py::arg("a"), py::arg("b"), "A custom docstring");
 
     {
         py::options options;
