@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from contextlib import redirect_stderr, redirect_stdout
 from io import StringIO
 
@@ -34,7 +36,7 @@ def test_captured_large_string(capsys):
 
 
 def test_captured_utf8_2byte_offset0(capsys):
-    msg = "\u07FF"
+    msg = "\u07ff"
     msg = "" + msg * (1024 // len(msg) + 1)
 
     m.captured_output_default(msg)
@@ -44,7 +46,7 @@ def test_captured_utf8_2byte_offset0(capsys):
 
 
 def test_captured_utf8_2byte_offset1(capsys):
-    msg = "\u07FF"
+    msg = "\u07ff"
     msg = "1" + msg * (1024 // len(msg) + 1)
 
     m.captured_output_default(msg)
@@ -54,7 +56,7 @@ def test_captured_utf8_2byte_offset1(capsys):
 
 
 def test_captured_utf8_3byte_offset0(capsys):
-    msg = "\uFFFF"
+    msg = "\uffff"
     msg = "" + msg * (1024 // len(msg) + 1)
 
     m.captured_output_default(msg)
@@ -64,7 +66,7 @@ def test_captured_utf8_3byte_offset0(capsys):
 
 
 def test_captured_utf8_3byte_offset1(capsys):
-    msg = "\uFFFF"
+    msg = "\uffff"
     msg = "1" + msg * (1024 // len(msg) + 1)
 
     m.captured_output_default(msg)
@@ -74,7 +76,7 @@ def test_captured_utf8_3byte_offset1(capsys):
 
 
 def test_captured_utf8_3byte_offset2(capsys):
-    msg = "\uFFFF"
+    msg = "\uffff"
     msg = "12" + msg * (1024 // len(msg) + 1)
 
     m.captured_output_default(msg)
@@ -84,7 +86,7 @@ def test_captured_utf8_3byte_offset2(capsys):
 
 
 def test_captured_utf8_4byte_offset0(capsys):
-    msg = "\U0010FFFF"
+    msg = "\U0010ffff"
     msg = "" + msg * (1024 // len(msg) + 1)
 
     m.captured_output_default(msg)
@@ -94,7 +96,7 @@ def test_captured_utf8_4byte_offset0(capsys):
 
 
 def test_captured_utf8_4byte_offset1(capsys):
-    msg = "\U0010FFFF"
+    msg = "\U0010ffff"
     msg = "1" + msg * (1024 // len(msg) + 1)
 
     m.captured_output_default(msg)
@@ -104,7 +106,7 @@ def test_captured_utf8_4byte_offset1(capsys):
 
 
 def test_captured_utf8_4byte_offset2(capsys):
-    msg = "\U0010FFFF"
+    msg = "\U0010ffff"
     msg = "12" + msg * (1024 // len(msg) + 1)
 
     m.captured_output_default(msg)
@@ -114,7 +116,7 @@ def test_captured_utf8_4byte_offset2(capsys):
 
 
 def test_captured_utf8_4byte_offset3(capsys):
-    msg = "\U0010FFFF"
+    msg = "\U0010ffff"
     msg = "123" + msg * (1024 // len(msg) + 1)
 
     m.captured_output_default(msg)
