@@ -20,7 +20,7 @@ namespace py = pybind11;
 namespace {
 
 struct IntStruct {
-    explicit IntStruct(int v) : value(v){};
+    explicit IntStruct(int v) : value(v) {};
     ~IntStruct() { value = -value; }
     IntStruct(const IntStruct &) = default;
     IntStruct &operator=(const IntStruct &) = default;
